requirements1.txt

keras>=2.3.1
imutils==0.5.3
numpy==1.18.2
opencv-python==4.2.0.*
pillow 
OS 
pygame
time
tkinter
tkinter.ttk

